Refactored code to take BigDecimal instead of Double value for any value that need calculation.

+++ backend/diet-generator/src/main/java/org/ectimel/dietgenerator/domain/model/Carbohydrates.java
@@ -1,14 +1,16 @@
 package org.ectimel.dietgenerator.domain.model;
 
-
 import lombok.AllArgsConstructor;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 @AllArgsConstructor
 public class Carbohydrates {
 
-    private Double totalCarbohydrates;
-    private Double fiber;
-    private Double sugar;
+    private BigDecimal totalCarbohydrates;
+    private BigDecimal fiber;
+    private BigDecimal sugar;
+
 }
