+++ backend/diet-generator/src/main/java/org/ectimel/dietgenerator/domain/model/Fats.java
@@ -3,9 +3,13 @@ package org.ectimel.dietgenerator.domain.model;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 public class Fats {
-    private Double totalFats;
-    private Double saturatedFats;
+    private BigDecimal totalFats;
+    private BigDecimal saturatedFats;
+
+
 }
